Update language for federal EV tax credit
## Links

- [Asana](https://app.asana.com/0/1208668890181682/1209198275861173)

## Description

To reflect that the credit is now deliverable as an upfront discount
at the point of sale (by being transferred to the dealership, I
think?). I removed the `(up to $7,500)` parenthetical from the
language in the ticket, to stay under the char limit, and because the
amount is in the `amount` field (which will be prominently shown in
the calculator).

Re: the ask to link to the DOE's `fueleconomy.gov` site instead: I
decided not to, because the [IRS page we currently link
to](https://www.irs.gov/credits-deductions/credits-for-new-clean-vehicles-purchased-in-2023-or-after)
is not bad, and it links to fueleconomy.gov itself. (In contrast,
fueleconomy.gov does not link to any IRS page as far as I can see.)

## Test Plan

Snapshot tests updated.