Fix import (#672)
## Description

The weekly HERO import didn't work, and it looks to be because of this
Husky post-checkout script. It turns out the newest version of Husky
runs hooks in `sh`, which does not support the double-paren-to-do-math
syntax that this hook was using (that's a Bash thing). Change it to be
sh-friendly.

## Test Plan

Will look at the CI logs on this PR to make sure there's no error
messages from this hook.