Calculate taxes for MA and NM (#675)
## Links

- [Asana](https://app.asana.com/0/1208668890181682/1209428558323182)

## Description

We have state tax credits for these, so we ought to calculate their
taxes.

I learned through doing this that a lot of MA tax calculators on the
internet (or at least the ones that have good SEO) are wrong; they
either don't account for the millionaire's tax or use the wrong
threshold for it.

I also learned that Wikipedia has the best SEO for "new mexico state
tax" but the table on the page is wrong; it doesn't include the top
marginal rate of 5.9%, which is new in 2021.

I'm left with a general feeling of irritation that I'm not quite sure
what to do with.

## Test Plan

New unit tests. The NM expected values are the ones given in the
source document. The MA values were calculated by me actually going
through MA tax forms and following the instructions.