Make filing status optional in v1 (#683)
## Links

- [Asana](https://app.asana.com/0/1208668890181682/1208959599500050)

## Description

If it's not passed, don't filter out tax credits (as we do if tax owed
is zero), and assume we don't meet filing-status-dependent low-income
thresholds.

Also don't cap the tax credit savings total at tax owed. This is a
non-issue in practice: the number is only shown in API v0, which
still requires filing status to be passed.

## Test Plan

Unit tests