Add item values for cool roofs and solar water heaters (#699)
## Links

-
[Asana](https://app.asana.com/1/1200412452784804/project/1208668890181682/task/1208676666376925)

## Description

This is the first step: allow API JSON to include these new items. The
next steps are:

- Add them to HERO so we can create actual incentives for these items.
- Add them to the embed so it can render them.
- Update the [public
docs](https://api.rewiringamerica.org/docs/routes/items) to describe
what these items are.

We've documented the `items` field in the API to say that we're going
to add values to the enum without considering it a breaking change, so
formally we shouldn't need to worry about backward compatibility. The
embed will handle this gracefully; it won't show incentives that are
for items it doesn't know about.

## Test Plan

`yarn test`. Manually change some incentives in JSON to have these
items and make sure tests still pass.