Fix comma parsing in query params
## Links

- [Asana](https://app.asana.com/0/1208668890181682/1209174180341578)

## Description

We've had it in our docs for a while that if you're passing multiple
values to the array-valued query params `authority_types` and `items`,
you can pass a comma-separated string: `authority_types=federal,state`.

However, that doesn't work in production, because Zuplo
percent-encodes commas in param values, so the backend would have seen
`authority_types=federal%2Cstate`, and our parsing library qs doesn't
split that into an array. (This is intentional in qs, not a bug.)

This went unnoticed for a long time because (a) it doesn't happen in
local development and (b) the calculator, which passes `items`, does
so by passing the same param multiple times (`items=xyz&items=abc`).

This fixes that problem by hooking into qs's custom decoding logic.
Although it may seem brittle to have a "decoder" function (which is
more meant for dealing with text encoding/decoding) doing this, it's
actually how qs works internally, so I'm not too worried.

The most brittle part of this is the hardcoded param names, which
apply the array-parsing logic on all endpoints. (The
`querystringParser` function isn't told what endpoint is being hit.)

In the course of making this change, I realized that Zuplo's
percent-encoding has actually been _covering up_ a potential bug of
ours: we would have parsed the `address` field as an array if it
contained un-encoded commas. This change solves that too.

## Test Plan

New unit test.

Use Bruno to hit localhost with encoded and un-encoded commas in the
`items` and `authority_types` field; make sure both work and produce
correspondingly filtered results. Put an address with commas (both
encoded and un-encoded) in the `address` field, and make sure that's
parsed correctly too.