Don't index low-income thresholds by state in JSON
## Description

The main point is to allow a set of thresholds not to belong to a
state at all (e.g. federal poverty level-based ones). But this also
unlocks a good bit of code cleanup.

This also removes the list of incentives in each threshold. In this
codebase it's entirely redundant -- the list of incentives is ignored
at runtime. It was originally there to support some spreadsheet-based
workflows.

## Test Plan

Unit tests.