Treat IRA incentives like state incentives
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-711)

## Description

Ever since state-level incentives were introduced, they've been
treated differently from IRA incentives: a different data type, fed
into an entirely different function to filter them for eligibility.

This fixes that situation: IRA incentives are now of the
`StateIncentive` type, and there's a single function that evaluates
eligibility for all incentives. This allows for a lot of code
simplification.

Another advantage is that we can represent the income qualifications
of the EV credits using the same "low income thresholds" logic as
state incentives, instead of the awkward situation of having copies of
a single incentive that differ only in the `agi_max_limit` field or
whatever. I added the 30D and 25E brackets in HERO a while back, and
use them here.

Since only the IRA tax credits were ever returned in v1, and not the
rebates, I deleted all the rebate from the JSON.

(I know all the IRA incentives are ending soon, and I could have just
waited and deleted all the "federal" code at once, but look... I
wanted a little treat _now_, and deleting thousands of lines of code
felt like just the thing.)

## Test Plan

`yarn test`.