Geo groups contain geographies now
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-769)

## Description

This is	to adapt to the	changes	in rewiringamerica/incentive_admin#328.

This refactors geo groups to contain geographies instead of
authorities. This greatly simplifies the geographic eligibility check.

I also decided to stop keeping a list of incentives with each geo
group. That has had no purpose in this codebase for quite a while now.

The one wrinkle was the Mass Save rule. Since it's no longer possible
to easily get a list of "Mass Save authorities" from the geo group, I
just hardcoded them. There is a unit test that will fail if those
hardcoded lists are ever out of sync with the contents of the
`ma-mass-save` and `ma-mass-save-electric` geo groups.

## Test Plan

`yarn test`.