Updates for new geography scheme
## Description

I've now added a geography or a geo group for every authority, and
removed redundant geo groups from incentives, and this is the
resulting update.

As a nice consequence, this now means (FINALLY) the programs API can
return programs from `other`-type authorities, because the geography
check is now the same for all authority types. The updated snapshot
proves this. Given that this was one of the top-level goals of this
whole geography refactor, this is pretty significant.

## Test Plan

`yarn test`