Revert to previous semantics of authority+geogroup location matching (#819)
## Links

- [Jira](https://rewiringamerica.atlassian.net/browse/RAT-877)

## Description

In the recent geography overhaul, I changed the semantics of an
incentive having a geo group: I made it so that the geo group entirely
overrode the geographic rule inherited from the incentive's authority.

It turns out this prevents expressing some important use cases, like a
utility offering enhanced incentives only to residents of specific
areas. So this PR restores it to the way it was before: _both_ the
authority's rule and the incentive's geo group are applied.

I'm also not bothering with the `console.error`; there is a unit test
that catches the case it was warning about, and we don't monitor for
error messages anyway.

I've checked all existing incentives to make sure that either their
eligibility won't be affected by this change, or this change fixes
their eligibility:

- For each incentive with a geo group, I checked whether the
  geographies in the group are a subset of the authority's geography
  or geo group.

- This was true in most cases (either the authority covered an entire
  state and the incentive covered a few counties or cities, or the
  authority had a consortium-of-utilities group and the incentive
  covered a subset of those utilities). Where this wasn't true, it's
  the situation I mentioned above: a utility, or consortium, offered
  incentives only to residents of a few counties/cities. This PR fixes
  those cases.

I've also got a HERO PR with a UI wording tweak that reflects this
change. I'll also update the "HERO Instructions" doc.

## Test Plan

`yarn test`.