Convert to better-sqlite3 (#825)
## Description

The combination of sqlite libraries we've been using hasn't been
updated for a while. In particular, it seems `sqlite3` doesn't have a
prebuilt binary for Node 22, which breaks the Docker build: it needs
Python within the builder image to compile SQLite, and it's not there,
and it's a whole mess.

This PR switches everything to use `better-sqlite3` instead: it's
actively maintained, and claims various advantages, which you can read
about [in their readme](https://github.com/WiseLibs/better-sqlite3).

You'll notice that the new lib is synchronous. The rationale for this
is explained in the better-sqlite3 docs, but as additional reassurance
for this context, our queries are so fast anyway and our traffic is so
low that there's not going to be any noticeable difference.

## Test Plan

`yarn test`

`yarn dev` and make some `/utilities` and `/calculator` requests with
Bruno; this exercises all of the SQLite queries.

`docker build .` to make sure this fixes the Docker build problem
caused by #824.

Ran the Docker image locally and made requests against it with Bruno
to make sure the built image is functional.